+++ vec/ChangeLog.md
@@ -1,5 +1,11 @@
 # Revision history for boring
 
+## 0.1.1
+
+- Add `Data.Vec.DataFamily.SpineStrict` module
+- Add `Data.Vec.DataFamily.SpineStrict.Pigeonhole` module:
+  this let us define `Representable` in a handy way.
+
 ## 0.1
 
 - Reverse dependencies with `boring`.